+++ tsdat/const/const.py
@@ -1,2 +1,6 @@
+COORDINATE_SYSTEM = "coord_sys"
+INPUT_DATASTREAM = "input_ds"
+OUTPUT_DATASTREAM = "output_ds"
+
 InputKey = str
 VarName = str
