+++ tsdat/transform/adi/adi_transformation_types.py
@@ -1,0 +1,7 @@
+class ADITransformationTypes:
+    # Allowed ADI transform algorithms - these are the values that can be used in transformation_type parameter
+    TRANS_AUTO = "TRANS_AUTO"
+    TRANS_INTERPOLATE = "TRANS_INTERPOLATE"
+    TRANS_SUBSAMPLE = "TRANS_SUBSAMPLE"
+    TRANS_BIN_AVERAGE = "TRANS_BIN_AVERAGE"
+    TRANS_PASSTHROUGH = "TRANS_PASSTHROUGH"
