+++ tsdat/transform/adi/bad_transformation_settings_error.py
@@ -1,0 +1,2 @@
+class BadTransformationSettingsError(ValueError):
+    """Exception class used when bad combinations of parameters are passed to transformation code."""
